Extract duplicate ticksOptions
