+++ app/javascript/charts/citizen-feedback/overall_rating_chart.js
@@ -1,16 +1,9 @@
 import GroupBarChart from '../group_bar_chart'
+import { ticksOptions } from '../../utils/bar_chart'
 
 class OverallFeedbackChart extends GroupBarChart {
   chartId = 'chart_overall_rating_by_owso';
-  ticksOptions = {
-    maxRotation: 45,
-    minRotation: 45,
-    callback: function(value) {
-      let maxLength = 10,
-          ellipsisValue = `${value.substr(0, 10)}...`;
-      return (value.length >= maxLength) ? ellipsisValue : value;
-    }
-  }
+  ticksOptions = ticksOptions
 
   format = () => {
     let { labels, dataset } = this.ds;
