+++ app/javascript/utils/bar_chart/index.js
@@ -19,3 +19,13 @@ export const extractBarDataset = (ds) => {
 export const suggestedMax = function ( array, scale = 1.2 ) {
   return _.max(array) * scale
 }
+
+export const ticksOptions = {
+  maxRotation: 45,
+  minRotation: 45,
+  callback: function(value) {
+    let maxLength = 10,
+        ellipsisValue = `${value.substr(0, 10)}...`;
+    return (value.length >= maxLength) ? ellipsisValue : value;
+  }
+}
